+++ api/config.js
@@ -19,11 +19,21 @@ var config = {
     oidc: {
       cookieKeys: [process.env.COOKIE_KEY, process.env.OLD_COOKIE_KEY],
       initialAccessToken: process.env.OIDC_INITIAL_ACCESS_TOKEN,
+      pairwiseSalt: process.env.OIDC_PAIRWISE_SALT,
     },
     baseUrl: process.env.OIDC_BASE_URL,
     aws: {
       s3Bucket: process.env.OIDC_S3_BUCKET,
     },
+    clientInitiatedLogout: process.env.CLIENT_INITIATED_LOGOUT === 'true' ? true : false,
+    userRegistration: process.env.ENABLE_USER_REGISTRATION === 'true' ? true : false,
+    webhooks: process.env.ENABLE_WEBHOOKS === 'true' ? {
+      adapter: process.env.WEBHOOK_ADAPTER ? process.env.WEBHOOK_ADAPTER : 'memory',
+      timeout: process.env.WEBHOOK_TIMEOUT || 2000,
+      maxRetries: process.env.WEBHOOK_MAX_RETRIES || 2,
+      retryDelay: process.env.WEBHOOK_RETRY_DELAY || 10000,
+      concurrency: process.env.WEBHOOK_CONCURRENCY || 1,
+    } : false,
   },
   'development': {},
   'qa': {},