+++ api/manifest.js
@@ -1,9 +1,12 @@
+const handlebars = require('handlebars');
+const ioc = require('electrolyte');
+const GoodWinston = require('good-winston');
+
 const config = require('./config');
 const formatError = require('./src/lib/format-error');
 const fetchKeystore = require('./src/lib/fetch-keystore');
-const handlebars = require('handlebars');
+const logger = require('./src/lib/logger');
 
-var ioc = require('electrolyte');
 ioc.use(ioc.dir('src/lib'));
 ioc.use(ioc.dir('src/application'));
 
@@ -62,14 +65,15 @@ module.exports = Promise.all([
       },
       {
         plugin: {
+          register: './plugins/email-token-scheme',
+        }
+      },
+      {
+        plugin: {
           register: 'good',
           options: {
             reporters: {
-              consoleReporter: [
-                {module: 'good-squeeze', name: 'Squeeze', args: [{log: '*', response: '*', error: '*', request: '*'}]},
-                {module: 'good-console'},
-                'stdout'
-              ]
+              winston: [ new GoodWinston({ winston: logger }) ]
             }
           }
         },
@@ -78,6 +82,7 @@ module.exports = Promise.all([
         plugin: {
           register: './plugins/openid-connect/openid-connect',
           options: {
+            logger,
             vision: {
               engines: {
                 hbs: handlebars
@@ -88,9 +93,11 @@ module.exports = Promise.all([
               layout: 'default',
             },
             prefix: 'op',
-            getTemplate: lib.themeService.renderThemedTemplate,
+            getTemplate: lib.themeService.renderThemedTemplate.bind(lib.themeService),
             authenticateUser: lib.userService.authenticate,
             findUserById: lib.userService.findByIdWithCtx,
+            userRegistration: config('userRegistration'),
+            pairwiseSalt: config('/oidc/pairwiseSalt'),
             cookieKeys: config('/oidc/cookieKeys'),
             initialAccessToken: config('/oidc/initialAccessToken'),
             adapter: function OidcAdapterFactory(name) {