+++ api/migrations/20171026204217_template_meta.js
@@ -1,0 +1,5 @@
+exports.up = knex => knex.schema.table('SIP_template', t => {
+  t.json('options');
+});
+
+exports.down = () => {/*noop*/};