+++ api/src/application/client/client-routes.js
@@ -1,8 +1,0 @@
-module.exports = (bookshelf) => {
-  return [];
-};
-
-module.exports['@singleton'] = true;
-module.exports['@require'] = [
-  'bookshelf',
-];