+++ api/src/application/email-token/email-token-model.js
@@ -1,0 +1,4 @@
+module.exports = bookshelf => ({
+  tableName: 'SIP_email_token',
+  idAttribute: 'token',
+});