+++ api/src/application/email/drivers/ses.js
@@ -2,24 +2,25 @@ const AWS = require('aws-sdk');
 AWS.config.update({region: 'us-west-2'});
 const ses = new AWS.SES({apiVersion: '2010-12-01'});
 const checkWhitelist = require('../check-whitelist');
+const Boom = require('boom');
 
 class SesDriver {
   send(emailObject) {
     return new Promise((resolve, reject) => {
       if (!emailObject.to) {
-        return reject('no to address provided');
+        return reject(new Error('no to address provided'));
       }
 
       if (!emailObject.subject) {
-        return reject('no subject provided');
+        return reject(new Error('no subject provided'));
       }
 
       if (!emailObject.text && !emailObject.html) {
-        return reject('no text or html body provided');
+        return reject(new Error('no text or html body provided'));
       }
 
       if (emailObject.attachments) {
-        return reject('ses driver does not currently support attachments');
+        return reject(new Error('ses driver does not currently support attachments'));
       }
 
       const params = {
@@ -38,11 +39,10 @@ class SesDriver {
       };
 
       ses.sendEmail(params).promise().then(result => {
-        console.log(result);
         resolve(result);
       })
         .catch(err => {
-          reject(err);
+          reject(Boom.wrap(err, err.statusCode));
         });
     });
   }