+++ api/src/application/user/user-oidc-service.js
@@ -1,18 +1,7 @@
-const bcrypt = require('bcrypt');
+const comparePasswords = require('../../lib/comparePasswords');
+const bookshelf = require('../../lib/bookshelf');
 
-module.exports = (bookshelf) => {
-  const comparePasswords = (password, user) => {
-    var hash = user.get('password');
-    return new Promise(function(resolve, reject) {
-      bcrypt.compare(password, hash, function(err, res) {
-        if (res) {
-          resolve(true);
-        } else {
-          resolve(false);
-        }
-      });
-    });
-  };
+module.exports = () => {
 
   const findByIdForOidc = id => bookshelf.model('user').where({ id }).fetch()
     .then(user => user ? user.serialize({ strictOidc: true }) : false);
@@ -34,9 +23,9 @@ module.exports = (bookshelf) => {
 
     findByIdWithCtx: function(ctx, id) {
       return findByIdForOidc(id);
-    }
+    },
   };
 };
 
 module.exports['@singleton'] = true;
-module.exports['@require'] = ['bookshelf'];
+module.exports['@require'] = [];